Fix the build for Python < 2.5.0
Fix (old) commit [9297] where Py_ssize_t was set to ssize_t for Python
versions anterior to 2.5.0 whereas it was an int.  We therefore had an
incompatible pointer type when calling PyDict_Next().

As INN is supposed to work on 2.2.0, this bug is now fixed.

Also use "0" and "1" instead of "False" and "True" that did not exist
in 2.2.0 for the distutils module.

+++ nnrpd/python.c
@@ -26,7 +26,7 @@
 /*  Python redefines _POSIX_C_SOURCE, so undef it to suppress warnings. */
 #undef _POSIX_C_SOURCE
 
-/*  Make "s#" use Py_ssize_t rather than signed size_t. */
+/*  Make "s#" use Py_ssize_t rather than int. */
 #define PY_SSIZE_T_CLEAN
 
 /*  Python.h must be included after having defined PY_SSIZE_T_CLEAN,
@@ -37,7 +37,7 @@
 
 /*  Define Py_ssize_t when it does not exist (Python < 2.5.0). */
 #if PY_VERSION_HEX < 0x02050000
-  typedef ssize_t Py_ssize_t;
+  typedef int Py_ssize_t;
 #endif
 
 #include "clibrary.h"
